renamed addCar to AddProductItemToCar, fixed dataArrButtons, fixed images
