+++ src/Components/Product.jsx
@@ -6,7 +6,7 @@ function Product({
   currencyFormat,
   price,
   installments,
-  onAddCar,
+  onAddProductItemToCar,
   completeProduct,
 }) {
   const [hovered, setHovered] = useState(false);
@@ -18,7 +18,7 @@ function Product({
       onMouseLeave={() => setHovered(false)}
     >
       <img
-        src={`/src/assets/products/${sku}-${hovered ? 2 : 1}-product.webp`}
+        src={`/../assets/products/${sku}-${hovered ? 2 : 1}-product.webp`}
         alt={sku}
       />
       <p className="text-xl">{title}</p>
@@ -30,7 +30,7 @@ function Product({
         className={`w-full h-14 text-white  ${
           hovered ? "bg-yellow-400" : "bg-gray-800"
         }`}
-        onClick={() => onAddCar(completeProduct)}
+        onClick={() => onAddProductItemToCar(completeProduct)}
       >
         Add to cart
       </button>
